Adding logic for retrieving current user
