+++ effectly/src/hooks/global/useCurrentUser.tsx
@@ -2,13 +2,15 @@ import { useEffect } from "react";
 import { auth } from "../../firebase/firebase";
 import { useNavigate } from "react-router-dom";
 import { onAuthStateChanged } from "firebase/auth";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { login, logout } from "../../store/users/userSlice";
+import { User, UserRootState } from "../../store/users/types";
 
 const useCurrentUser = () => {
   const navigate = useNavigate();
   const dispatchToStore = useDispatch();
 
+  // Redirect when user is not logged in - otherwise, add him to global store
   useEffect(() => {
     onAuthStateChanged(auth, (user) => {
       if (user !== null) {
@@ -20,6 +22,11 @@ const useCurrentUser = () => {
       }
     });
   }, []);
+
+  // Retrieve the currently signed in user from the store
+  const { user } = useSelector<UserRootState, User>((state) => state.user);
+
+  return user;
 };
 
 export default useCurrentUser;
