+++ effectly/src/store/users/types.ts
@@ -1,0 +1,7 @@
+export interface User {
+  user: any;
+}
+
+export interface UserRootState {
+  user: User;
+}
